separate gulag api out, setup b.ppy.sh, change routing, more
