--- app/__init__.py
@@ -1,7 +1,0 @@
-from . import api
-from . import bg_loops
-from . import constants
-from . import misc
-from . import objects
-from . import services
-from . import settings
