+++ app/api/__init__.py
@@ -1,0 +1,2 @@
+from . import domains  # folders
+from . import init_api  # files
