+++ app/api/domains/map.py
@@ -1,22 +1,17 @@
-import re
-from typing import Optional
-from typing import Union
-
-from cmyui.web import Connection
-from cmyui.web import Domain
-
-from app.objects import glob
-
-HTTPResponse = Optional[Union[bytes, tuple[int, bytes]]]
-
 """ bmap: static beatmap info (thumbnails, previews, etc.) """
+from fastapi import APIRouter
+from fastapi import status
+from fastapi.requests import Request
+from fastapi.responses import RedirectResponse
 
-BASE_DOMAIN = glob.config.domain
-domain = Domain({f"b.{BASE_DOMAIN}", "b.ppy.sh"})
+
+router = APIRouter(prefix="/map", tags=["Beatmaps"])
 
 # for now, just send everything to osu!
 # eventually if we do bmap submission, we'll need this.
-@domain.route(re.compile(r"^.+$"))
-async def everything(conn: Connection) -> HTTPResponse:
-    conn.resp_headers["Location"] = f"https://b.ppy.sh{conn.path}"
-    return (301, b"")
+@router.get("/{file_path:path}")
+async def everything(request: Request):
+    return RedirectResponse(
+        url=f"https://b.ppy.sh{request['path']}",
+        status_code=status.HTTP_301_MOVED_PERMANENTLY,
+    )
