--- app/misc/__init__.py
@@ -1,2 +1,0 @@
-from . import context
-from . import utils
